Bug fix in the sign of the indexes of FOR_ALL_ELEMENTS_IN_ARRAY3D
git-svn-id: file:///home/roberto/Git/svn@5794 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 3357f0d59ff79139e43b2e3d62234f20791ddfb3

+++ trunk/xmipp/libraries/data/multidim_array.h
@@ -277,9 +277,9 @@ extern std::string floatToString(float F, int _width, int _prec);
  * @endcode
  */
 #define FOR_ALL_ELEMENTS_IN_ARRAY3D(V) \
-    for (size_t k=STARTINGZ(V); k<=FINISHINGZ(V); k++) \
-        for (size_t i=STARTINGY(V); i<=FINISHINGY(V); i++) \
-            for (size_t j=STARTINGX(V); j<=FINISHINGX(V); j++)
+    for (int k=STARTINGZ(V); k<=FINISHINGZ(V); k++) \
+        for (int i=STARTINGY(V); i<=FINISHINGY(V); i++) \
+            for (int j=STARTINGX(V); j<=FINISHINGX(V); j++)
 
 /** For all elements in common.
  * @ingroup MultidimArraySizeShape
